submit: drop a redundant cast on an already-uint8_t field
record.holdings_kind is uint8_t (shekyl_types.h:1179) and is passed
unwrapped to shekyl_archival_last_served_scan three lines above, so the
static_cast asserted a conversion that was not happening. A cast that
narrows nothing reads as a reminder that the source type is wider, which
is the wrong thing to leave next to a discriminant.

+++ src/rpc/daemon_submit_ffi.cpp
@@ -226,7 +226,7 @@ bool fill_unbond_facts_locked(Blockchain& bc, const crypto::hash& p_id,
   out.bond_spend_pk =
     handle.bond_spend_pk.empty() ? nullptr : handle.bond_spend_pk.data();
   out.bond_spend_pk_len = handle.bond_spend_pk.size();
-  out.holdings_kind = static_cast<uint8_t>(record.holdings_kind);
+  out.holdings_kind = record.holdings_kind;
   out.last_served_scan = scan;
   out.per_shard_last_served =
     handle.per_shard_last_served.empty() ? nullptr : handle.per_shard_last_served.data();
